# PctBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PCT UVCI — API REST",
        "description": "API de gestion des activités pédagogiques et volumes horaires — UVCI",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/annees": {
            "get": {
                "tags": [
                    "Années"
                ],
                "summary": "Liste des années académiques",
                "operationId": "53c13037109d4ebe427702d98f0e3f00",
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Années"
                ],
                "summary": "Créer une année",
                "operationId": "3fc0cd0f616dcc2a60542c67c578e91c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle_annee",
                                    "date_debut",
                                    "date_fin"
                                ],
                                "properties": {
                                    "libelle_annee": {
                                        "type": "string",
                                        "example": "2025-2026"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/annees/{id}": {
            "get": {
                "tags": [
                    "Années"
                ],
                "summary": "Détail",
                "operationId": "f002ab85d595090062d4ed4151bebc8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Année"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Années"
                ],
                "summary": "Modifier",
                "operationId": "036d406a395d580803f29c0bacfc181a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle_annee": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Années"
                ],
                "summary": "Supprimer",
                "operationId": "2ac3f15c53bc575282e0779dc2b44540",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/annees/{id}/activer": {
            "patch": {
                "tags": [
                    "Années"
                ],
                "summary": "Définir comme année active",
                "operationId": "31f0bd9263fcf4aba03ba8e27ef4fced",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/cours": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Liste paginée des cours",
                "operationId": "c183775c7c33648dbbd255db06aac9d3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cours"
                ],
                "summary": "Créer un cours",
                "operationId": "41554cbc84ede3082c5201915b2f2f09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intitule_ecue",
                                    "niveau",
                                    "semestre",
                                    "credit_ecue",
                                    "charge_horaire_annuel"
                                ],
                                "properties": {
                                    "intitule_ecue": {
                                        "type": "string"
                                    },
                                    "niveau": {
                                        "type": "string",
                                        "enum": [
                                            "L1",
                                            "L2",
                                            "L3",
                                            "M1",
                                            "M2"
                                        ]
                                    },
                                    "semestre": {
                                        "type": "string",
                                        "example": "S1"
                                    },
                                    "credit_ecue": {
                                        "type": "integer"
                                    },
                                    "charge_horaire_annuel": {
                                        "type": "integer"
                                    },
                                    "code_specialite": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cours créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cours/{id}": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Détail",
                "operationId": "c92bc269c93fbedc8d5a5323b1a55fb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cours"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cours"
                ],
                "summary": "Modifier",
                "operationId": "5f31672184168033f10b1e426749d763",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intitule_ecue": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cours"
                ],
                "summary": "Supprimer",
                "operationId": "b0ec197ca405514af0ede84b7ffce25c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token obtenu via POST /login. Format : Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}